broke up map_selection_live.ex into two files
